clean-project/Integrated_Model/requirements.txt
streamlit
tensorflow
tensorflow-macos
pillow
"numpy<2.0"
gradio
torch 
torchvision
torchaudio
transformers
peft
flask 
requests
